refactor: removes ApiLayer
