+++ app/src/main/java/com/omouravictor/ratesnow/data/repository/BitCoinRepository.kt
@@ -3,7 +3,7 @@ package com.omouravictor.ratesnow.data.repository
 import com.omouravictor.ratesnow.data.local.AppDataBase
 import com.omouravictor.ratesnow.data.local.entity.BitCoinEntity
 import com.omouravictor.ratesnow.data.network.hgbrasil.bitcoin.SourceRequestBitcoinModel
-import com.omouravictor.ratesnow.data.network.hgbrasil.rates.ApiService
+import com.omouravictor.ratesnow.data.network.ApiService
 import com.omouravictor.ratesnow.utils.Resource
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
