+++ app/src/main/java/com/omouravictor/ratesnow/data/repository/RatesRepository.kt
@@ -3,7 +3,7 @@ package com.omouravictor.ratesnow.data.repository
 import com.omouravictor.ratesnow.data.local.AppDataBase
 import com.omouravictor.ratesnow.data.local.entity.CurrencyEntity
 import com.omouravictor.ratesnow.data.network.base.ResultStatus
-import com.omouravictor.ratesnow.data.network.hgbrasil.rates.ApiService
+import com.omouravictor.ratesnow.data.network.ApiService
 import com.omouravictor.ratesnow.data.network.hgbrasil.rates.SourceRequestCurrencyModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
