+++ app/src/main/java/com/omouravictor/ratesnow/data/repository/StocksRepository.kt
@@ -2,7 +2,7 @@ package com.omouravictor.ratesnow.data.repository
 
 import com.omouravictor.ratesnow.data.local.AppDataBase
 import com.omouravictor.ratesnow.data.local.entity.StockEntity
-import com.omouravictor.ratesnow.data.network.hgbrasil.rates.ApiService
+import com.omouravictor.ratesnow.data.network.ApiService
 import com.omouravictor.ratesnow.data.network.hgbrasil.stock.SourceRequestStockModel
 import com.omouravictor.ratesnow.utils.Resource
 import kotlinx.coroutines.flow.Flow
